move ISR of uart 0 to own c file
save flash when serial is not used.
