+++ ch55xduino/ch55x/cores/ch55xduino/HardwareSerial.h
@@ -10,6 +10,8 @@
 #define SERIAL1_TX_BUFFER_SIZE 16
 #define SERIAL1_RX_BUFFER_SIZE 16
 
+#define UART0_FLG_SENDING     (1<<0)
+
 uint8_t Serial0(void);
 void Serial0_begin(unsigned long baud);
